+++ docs/sidebars.js
@@ -60,6 +60,11 @@ module.exports = {
           'rcv_api/addAnchors',
         ],
       },
+      {
+        type: 'category',
+        label: 'Rich Content Preview [legacy]',
+        items: ['rcp_api/RichContentPreviewAPI'],
+      },
     ],
   },
   devs: {
@@ -76,6 +81,6 @@ module.exports = {
       'ricos/adding-a-viewer',
       'ricos/migrating-from-rich-content',
     ],
-    Features: ['ricos/theming'],
+    Features: ['ricos/theming', 'ricos/preview'],
   },
 };