+++ examples/storybook/stories/Components/ViewerWrapper.js
@@ -77,9 +77,16 @@ const ViewerWrapper = ({
   isMobile = mobileDetect.mobile() !== null,
   addAnchors,
   normalize,
+  preview,
 }) => {
   return (
-    <RicosViewer plugins={plugins} theme={{ palette }} content={content} isMobile={isMobile}>
+    <RicosViewer
+      plugins={plugins}
+      theme={{ palette }}
+      content={content}
+      isMobile={isMobile}
+      preview={preview}
+    >
       <RichContentViewer addAnchors={addAnchors} normalize={normalize} />
     </RicosViewer>
   );
@@ -89,6 +96,7 @@ ViewerWrapper.propTypes = {
   content: PropTypes.object,
   palette: PropTypes.arrayOf(PropTypes.object),
   isMobile: PropTypes.bool,
+  preview: PropTypes.object,
   addAnchors: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
   normalize: PropTypes.object,
   _rcProps: PropTypes.object,