+++ packages/preview/web/src/ContentStateBuilder/ContentStateBuilder.js
@@ -1,14 +1,12 @@
-import { Version, RicosContent } from 'wix-rich-content-common';
+import { Version } from 'wix-rich-content-common';
 import { METHOD_BLOCK_MAP, METHOD_GROUPED_BLOCK_MAP, METHOD_PLUGIN_DATA_MAP } from '../const';
 import { toArray, mergeBlockWithEntities, addPlugin } from './builder-utils';
 import { readMore, seeFullPost, imageCounter } from '../Interactions/interaction-utils';
 
-const DEFAULT_STATE: RicosContent = { blocks: [], entityMap: {}, VERSION: Version.currentVersion };
+const DEFAULT_STATE = { blocks: [], entityMap: {}, VERSION: Version.currentVersion };
 
 class ContentStateBuilder {
-  contentState: RicosContent;
-
-  constructor(initialState: RicosContent) {
+  constructor(initialState) {
     this.contentState = { ...DEFAULT_STATE, ...(initialState || {}) };
   }