+++ packages/preview/web/src/index.ts
@@ -1,3 +1,15 @@
+import { interactionMap } from './Interactions/interactionMap';
+import { defaultTransformation } from './Components/default-transformation';
 export { default as ContentStateTransformation } from './RuleEngine/ContentStateTransformation';
-export { interactionMap } from './Interactions/interactionMap';
 export { default as RichContentPreview } from './Components/RichContentPreview';
+export { interactionMap };
+
+const defaultConfig = {
+  transformation: defaultTransformation,
+  contentInteractionMappers: [interactionMap],
+};
+
+export const previewSettings = (config = {}) => ({
+  ...defaultConfig,
+  ...config,
+});