+++ packages/ricos-common/web/src/types.ts
@@ -1,3 +1,4 @@
+export * from './previewStrategy/previewTypes';
 export * from './pluginsStrategy/pluginTypes';
 export * from './themeStrategy/themeTypes';
 export * from './RicosTypes';